Shellcheck compliance: SC1090 fixes
SC1090: Can't follow non-constant source. Use a directive to specify
location.

The recently added pre-commit hook sources a file, but within a
conditional statement, since not all developers will store their
virtualenv in the same location. Shellcheck isn't smart enough to parse
the conditional, so just ignore the implied source path.

If we use the shellcheck directive to point to the path, and it doesn't
exist, shellcheck will still fail, since it can't read in the file
at the designated source path. The shellcheck wiki recommends the
/dev/null approach in such scenarios.

URL to the Shellcheck wiki for this rule:

https://github.com/koalaman/shellcheck/wiki/SC1090